+++ src/devToolsExtensionContext.ts
@@ -22,6 +22,8 @@ import { DiagnosticsService } from './services/diagnosticsService';
 import { ALImagesService } from './services/alImagesService';
 import { DuplicateCodeService } from './services/duplicateCodeService';
 import { WarningDirectivesService } from './services/warningDirectivesService';
+import { HoverService } from './services/hoverService';
+import { ReferencesService } from './services/referencesService';
 
 export class DevToolsExtensionContext implements vscode.Disposable {
     alLangProxy : ALLangServerProxy;    
@@ -43,6 +45,8 @@ export class DevToolsExtensionContext implements vscode.Disposable {
     alImagesService: ALImagesService;
     duplicateCodeService: DuplicateCodeService;
     warningDirectivesService: WarningDirectivesService;
+    hoverService: HoverService;
+    referencesService: ReferencesService;
 
     constructor(context : vscode.ExtensionContext) {
         this.alLangProxy = new ALLangServerProxy()
@@ -70,6 +74,8 @@ export class DevToolsExtensionContext implements vscode.Disposable {
         this.alImagesService = new ALImagesService(this);
         this.duplicateCodeService = new DuplicateCodeService(this);
         this.warningDirectivesService = new WarningDirectivesService(this);
+        this.hoverService = new HoverService(this);
+        this.referencesService = new ReferencesService(this);
     }
 
     getUseSymbolsBrowser() : boolean {