+++ src/langserver/toolsReferencesResponse.ts
@@ -1,0 +1,5 @@
+import { DocumentTextRange } from "../symbollibraries/documentTextRange";
+
+export class ToolsReferencesResponse {
+    references?: DocumentTextRange[];
+}